% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_manage_dependencies.R
\name{add_dependencies_to_description}
\alias{add_dependencies_to_description}
\title{add_dependencies_to_description() is deprecated}
\usage{
add_dependencies_to_description(...)
}
\arguments{
\item{...}{nothing}
}
\description{
Please use \code{attachment::att_to_description()} from https://cran.r-project.org/package=attachment.
}
