% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.AF.summary.r
\name{calculate.AF.summary}
\alias{calculate.AF.summary}
\title{Returns a list of allele frequencies and heterozygosities}
\usage{
calculate.AF.summary(dart_data)
}
\arguments{
\item{dart_data}{-- a dart data object [Required]}
}
\value{
a list containing a vector of alternate allele frequencies
and a vector of observed heterozygosities
}
\description{
Input is a dart data object with altcount genotype encoding
(0=hom ref, 1=het, 2=hom alt).
}
\examples{
AF.summary <- calculate.AF.summary(dart_data)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
