% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.genome.LD.decay.r
\name{calculate.genome.LD.decay}
\alias{calculate.genome.LD.decay}
\title{Returns estimates of Linkage}
\usage{
calculate.genome.LD.decay(dart_data, MAF_thresh = 0.1,
  min_dist_thresh_bp = 150, max_dist_thresh_bp = 1e+07)
}
\arguments{
\item{dart_data}{-- a dart meta data object [Required]}
}
\value{
a list containing info about snp pairs
}
\description{
Input is a dart, meta, genome object
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

