% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.population.allele.stats.r
\name{calculate.population.allele.stats}
\alias{calculate.population.allele.stats}
\title{Calculates alternate allele counts and numbers of sampled alleles for  
a dart object. Results appropriate for BEDASSLE input}
\usage{
calculate.population.allele.stats(dart_data, population)
}
\arguments{
\item{dart_data}{-- a dart data object [Required]}

\item{pop}{-- a vector containing population assignments for each sample}
}
\value{
a list containing names of infiles for range expansion analysis
}
\description{
Input is a dart data object with altcount genotype encoding
(0=hom ref, 1=het, 2=hom alt).
}
\examples{
\dontrun{
popstats <- calculate.population.allele.stats(dart_data, pop)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

