% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_raster_to_points.r
\name{clip_raster_to_points}
\alias{clip_raster_to_points}
\title{Clip a raster to a convex hull surrounding a set of points}
\usage{
clip_raster_to_points(raster_to_clip, location_file = NULL, gdm_dir = NULL,
  hull_res = 150, convexity = 0.85)
}
\arguments{
\item{basedir}{-- name of the base directory for R&R}

\item{species}{-- species name}

\item{dataset}{-- dataset name}

\item{treatment}{}

\item{pop}{-- a vector of population assignments}

\item{Ksel}{-- the selected number of ancestral populations}
}
\value{
file name
}
\description{
Clip a raster to a convex hull surrounding a set of points
}
\examples{
\dontrun{
new_raster <- clip_raster_to_points(old_raster, location_file="locations.txt")
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
