% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_diversity_data.r
\name{collection_diversity_data}
\alias{collection_diversity_data}
\title{Calculates diversity (proportion of sites polymorphism) for
collections of sites}
\usage{
collection_diversity_data(dms, basedir, species, dataset, pop,
  sites_single = TRUE, sites_paired = TRUE)
}
\arguments{
\item{dms}{-- a dart data object [Required]}

\item{basedir}{-- name of the base directory for R&R}

\item{species}{-- species name}

\item{dataset}{-- dataset name}

\item{pop}{-- a vector containing population assignments for each sample}
}
\value{
a list containing names of infiles for range expansion analysis
}
\description{
Input is a dart data object with altcount genotype encoding
(0=hom ref, 1=het, 2=hom alt).
}
\examples{
\dontrun{
dart_gi <- dart2rangexp(dms, meta.csv)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
