% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_create_compendium.R
\name{create_compendium}
\alias{create_compendium}
\title{Quickly create a basic research compendium by combining several rrtools functions into one.}
\usage{
create_compendium(pkgname, data_in_git = TRUE)
}
\arguments{
\item{pkgname}{location to create new package. The last component of the path will be used as the package name}

\item{data_in_git}{should git track the files in the data directory? Default is TRUE}
}
\description{
In one step, this will create an R package, attach the MIT license to it, add the rrtools' README to it, initiate a Git repository and make an initial commit to track files in the package, and create the 'analysis' directory structure, and populate it with an R Markdown file and bib file. This function will not create a GitHub repository for the compendium, a Dockerfile, a Travis config file, or any package tests. Those require some interaction outside of R and are left to the user.
}
