% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{current_git_commit}
\alias{current_git_commit}
\title{Collect information about the last git commit for printing}
\usage{
current_git_commit(path)
}
\arguments{
\item{path}{path to the git repository. For example this will be
path = "../..", if we are in analysis/paper/paper.Rmd}
}
\value{
list
}
\description{
This looks into the local git repository for the project and
collects basic information about the last commit, such as the date, the SHA,
and the author of the commit. We can use this information
to print a sentence at the end of our document that allows us to identify the
exact commit that produced the rendered document output. This can be helpful
for knowing what is the most current version when there are multiple
versions of the rendered output in circulation among co-authors.
}
\examples{
\dontrun{
current_git_commit()
}
}
