% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.genome.hits.merge.r
\name{dart.genome.hits.merge}
\alias{dart.genome.hits.merge}
\title{Merges dart and meta data objects into a single object}
\usage{
dart.genome.hits.merge(dart_data, basedir, species, dataset)
}
\arguments{
\item{dart_data}{-- a dart data object [Required]}

\item{basedir}{-- directory for analysis [Required]}

\item{species}{-- the 8 letter species identifier [Required]}

\item{dataset}{-- name of dataset [Required]}
}
\value{
dart_data object with genome info fields
}
\description{
Input is a dart data object with altcount genotype encoding
(0=hom ref, 1=het, 2=hom alt).
}
\examples{
\dontrun{
dart_data <- dart.genome.hits.merge(dart_data, basedir, species, dataset)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

