% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.meta.data.merge.r
\name{dart.meta.data.merge}
\alias{dart.meta.data.merge}
\title{Merges dart and meta data objects into a single object}
\usage{
dart.meta.data.merge(dart_data, meta_data)
}
\arguments{
\item{dart_data}{-- a dart data object [Required]}

\item{meta_data}{-- a meta data object [Required]}
}
\value{
a dart_meta_data object
}
\description{
Input is a dart data object with altcount genotype encoding
(0=hom ref, 1=het, 2=hom alt).
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
