% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart2gds.r
\name{dart2gds}
\alias{dart2gds}
\title{Returns a dart data object as a genind object}
\usage{
dart2gds(dart_data, basedir, species, dataset, meta_data = FALSE)
}
\arguments{
\item{dart_data}{-- a dart data object [Required]}

\item{meta_data}{-- name of a file containing meta-data for samples}
}
\value{
a genind \{adegenet\} object containing the same data
}
\description{
Input is a dart data object with altcount genotype encoding
(0=hom ref, 1=het, 2=hom alt).
}
\examples{
\dontrun{
dart_gds <- dart2gds(dart_data, basedir, species, dataset, meta_data=FALSE)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
