% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart2JSFS.r
\name{dart2JSFS}
\alias{dart2JSFS}
\title{Generates the JSFS for two nominated populations}
\usage{
dart2JSFS(dms, basedir, species, samples)
}
\arguments{
\item{dms}{-- a dart data object [Required]}

\item{basedir}{-- name of the base directory for R&R}

\item{species}{-- species name}

\item{samples}{-- list containing indices for the populations for JSFS}

\item{dataset}{-- dataset name}
}
\value{
a list containing names of infiles for localdiff analysis
}
\description{
Input is a dart data object with altcount genotype encoding
(0=hom ref, 1=het, 2=hom alt).
}
\examples{
\dontrun{
ld_files <- dart2localdiff(gms, etc)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

