% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart2newhy.r
\name{dart2newhy}
\alias{dart2newhy}
\title{Writes a dart data object to a newhybrids input file}
\usage{
dart2newhy(dart_data, basedir, species, dataset, meta_data = FALSE)
}
\arguments{
\item{dart_data}{-- a dart data object [Required]}
}
\value{
an object
}
\description{
Input is a dart data object with altcount genotype encoding
(0=hom ref, 1=het, 2=hom alt).
}
\examples{
\dontrun{
dart_lfmm <- dart2lfmm(dart_data, meta=FALSE)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

