% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart2rangexp.r
\name{dart2rangexp}
\alias{dart2rangexp}
\title{Returns a dart data object in appropriate files for analysis 
with the Peter and Slatkin method of detecting a range expansion}
\usage{
dart2rangexp(dart_data, basedir, species, dataset, pop)
}
\arguments{
\item{dart_data}{-- a dart data object [Required]}

\item{basedir}{-- name of the base directory for R&R}

\item{species}{-- species name}

\item{dataset}{-- dataset name}

\item{pop}{-- a vector containing population assignments for each sample}
}
\value{
a list containing names of infiles for range expansion analysis
}
\description{
Input is a dart data object with altcount genotype encoding
(0=hom ref, 1=het, 2=hom alt).
}
\examples{
\dontrun{
dart_gi <- dart2rangexp(dart_data, meta.csv)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

