% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart2struct.r
\name{dart2struct}
\alias{dart2struct}
\title{Writes a dart data object to an lfmm genotype file}
\usage{
dart2struct(dart_data, basedir, species, dataset, use_pops = FALSE, pops)
}
\arguments{
\item{dart_data}{-- a dart data object [Required]}
}
\value{
an object
}
\description{
Input is a dart data object with altcount genotype encoding
(0=hom ref, 1=het, 2=hom alt).
}
\examples{
\dontrun{
dart_struct <- dart2struct(dart_data, pop_prior=FALSE)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
