% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.by.meta.fields.r
\name{data.by.meta.fields}
\alias{data.by.meta.fields}
\title{Merges dart and meta data objects into a single object}
\usage{
data.by.meta.fields(dart_data, fields, basedir, species, dataset,
  object = "files")
}
\arguments{
\item{dart_data}{-- a merged dart-meta data object [Required]}

\item{fields}{-- vector data field names for splitting [Required]}

\item{basedir}{-- name of the base directory [Required]}

\item{species}{-- species name [Required]}

\item{dataset}{-- dataset identifier [Required]}

\item{object}{-- if FALSE, return a vector of file names}
}
\value{
object   -- a list of file names for the data objects that are produced

field    -- the data object for the nominated field
}
\description{
Input is a dart data object with altcount genotype encoding
(0=hom ref, 1=het, 2=hom alt).
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
