% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_estimates_MOP.r
\name{diversity_estimates_MOP}
\alias{diversity_estimates_MOP}
\title{Provides an estimate of genetic diversity based on  
muliplicities of polymorphism occuring in DArT sequences}
\usage{
diversity_estimates_MOP(gt, method = "mean")
}
\arguments{
\item{gt}{-- genotype matrix [Required]}

\item{method}{-- can be "diags", "mean", "mle", or "abc"}

\item{sample}{-- sample size, in chromosomes}
}
\value{
a list containing diversity estimates
}
\description{
Input is a DArT style genotype matrix
}
\examples{
div_calcs <- diversity_estimates_MOP(gt, method="mean")
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

