% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.snps.r
\name{downsample.snps}
\alias{downsample.snps}
\title{Sample one SNP per locus at random}
\usage{
downsample.snps(dart_data, number = 1000, seed = 12345)
}
\arguments{
\item{dart_data}{-- dart data list  [required]}

\item{number}{-- number of random snps to select}

\item{seed}{-- random seed}
}
\description{
downsample.snps() receives data in list format, and samples a nominated number
of snp sites at random, with the treatment field updated to "downSamp"
}
\examples{
downsample.snps(dart_data, number=1000, seed=12345)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

