% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.dart2altcount.r
\name{encode.dart2altcount}
\alias{encode.dart2altcount}
\title{Recodes genotypes supplied from DArT format to count of alternate allele
Also changes the status from the $encoding element of dart data list 
from 'DArT' to 'altcount'}
\usage{
encode.dart2altcount(dart_data)
}
\arguments{
\item{dart_data}{-- dart data list  [required]}
}
\description{
Recodes genotypes supplied from DArT format to count of alternate allele
Also changes the status from the $encoding element of dart data list 
from 'DArT' to 'altcount'
}
\examples{
encode.dart2altcount(dart_data)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
