% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.samples.r
\name{exclude.samples}
\alias{exclude.samples}
\title{Remove a list of samples from a DArT data object  
Optionally removes markers that become monomorphic}
\usage{
exclude.samples(dart_data, excluded_sample_file, remove_fixed_loci = TRUE)
}
\arguments{
\item{dart_data}{-- dart data list  [required]}

\item{excluded_sample_file}{-- name of a file containing list of samples to be excluded [required]}

\item{removed_fixed_loci}{-- removed snps that are fixed after removal of nominated samples}
}
\description{
Remove a list of samples from a DArT data object  
Optionally removes markers that become monomorphic
}
\examples{
exclude.samples(dart_data, excluded_sample_file, remove_fixed_loci=TRUE)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

