% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.loci.MAC.lte.value.r
\name{find.loci.MAC.lte.value}
\alias{find.loci.MAC.lte.value}
\title{Find loci that have a minor allele count (MAC) lower than (or equal to) a nominated value
Note: if value is set to 0, it will find fixed (non-polymorphic) loci. If none found, returns "None".}
\usage{
find.loci.MAC.lte.value(dart_data, value = 0, return_as_names = FALSE)
}
\arguments{
\item{dart_data}{-- dart data list  [required]}

\item{value}{-- minimum value of minor allele count}

\item{return_as_names}{-- if TRUE, returns names of loci, if false, returns indices}
}
\description{
find.loci.MAC.lte.value() searches a dart data object for SNPs with MAC less than or equal to a nominated value
}
\examples{
find.loci.MAC.lte.value(dart_data, value=0, input_as_names=FALSE)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

