% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gdm.r
\name{format_gdm}
\alias{format_gdm}
\title{Prepares SNMF Q matrices to act as explanator variables in GDM}
\usage{
format_gdm(gdm_dir, Q = FALSE)
}
\arguments{
\item{gdm_dir}{-- name of directory containing GDM data}

\item{Q}{-- if TRUE, use Q matrix in GDM}
}
\value{
file name
}
\description{
Prepares SNMF Q matrices to act as explanator variables in GDM
}
\examples{
\dontrun{
gdm_object <- format_gdm(gdm_dir, Q=FALSE)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
