% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.genome.locus.pairs.r
\name{get.genome.locus.pairs}
\alias{get.genome.locus.pairs}
\title{Returns loci that are near each other on a eucalypt chromosome}
\usage{
get.genome.locus.pairs(dart_data, MAF_thresh, min_dist_thresh_bp = 150,
  max_dist_thresh_bp = 1e+07)
}
\arguments{
\item{dart_data}{-- a dart meta data object [Required]
MAF_thresh         -- minimum minor allele frequency
min_dist_thresh_bp -- minimum distance apart of loci
max_dist_thresh_bp -- maximum distance apart of loci}
}
\value{
a list containing info about snp pairs
}
\description{
Input is a dart, meta, genome object
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
