% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.intra.locus.pairs.r
\name{get.intra.locus.pairs}
\alias{get.intra.locus.pairs}
\title{Returns loci that are near each other on a eucalypt chromosome}
\usage{
get.intra.locus.pairs(dart_data, MAF_thresh = 0.2, min_dist_thresh_bp = 10)
}
\arguments{
\item{dart_data}{-- a dart meta data object [Required]}
}
\value{
a list containing info about snp pairs
}
\description{
Input is a dart, meta, genome object
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

