% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.population.info.r
\name{get.population.info}
\alias{get.population.info}
\title{Returns information about populations, extracted from meta data}
\usage{
get.population.info(m, p, method = "average")
}
\arguments{
\item{m}{-- a meta data object [Required]}

\item{p}{-- the population assignments [Required]}
}
\value{
a list containing locations, population names, population number
}
\description{
Input is a meta data object
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

