% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_data.r
\name{get_climate_data}
\alias{get_climate_data}
\title{Gets data for a list of climate data 
for a set of locations}
\usage{
get_climate_data(varfile, climdir, longlatpoints)
}
\arguments{
\item{climdir}{-- name of the directory containing gtif files}

\item{varlist}{-- A file with two columns, 'var' and 'file'
var lists the variable names
file is the name of the gtif}

\item{longlat}{-- a set of long and lat info for data points}
}
\value{
-- climate variable data for each location
}
\description{
Gets data for a list of climate data 
for a set of locations
}
\examples{
\dontrun{
VarClimLoc <- get_climate_data(varlist=RRStandardClimLayers, climdir="/home/jason/RandR/clim", longlat=long_lat.txt)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

