% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.pw.IBD.r
\name{individual.pw.IBD}
\alias{individual.pw.IBD}
\title{Returns a matrix of spatial distances among populations}
\usage{
individual.pw.IBD(dart_data, basedir, species, dataset)
}
\arguments{
\item{dart_data}{-- a dart meta data object [Required]}
}
\value{
a matrix containing spatial distances between populations
}
\description{
Input is a dart meta data object
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

