% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.pw.spatial.dist.r
\name{individual.pw.spatial.dist}
\alias{individual.pw.spatial.dist}
\title{Returns a matrix of spatial distances among individuals}
\usage{
individual.pw.spatial.dist(dart_data)
}
\arguments{
\item{dart_data}{-- a dart meta data object [Required]}
}
\value{
a matrix containing spatial distances between populations
}
\description{
Input is a dart meta data object
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
