% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_rrtools}
\alias{init_rrtools}
\title{writes .Rprofile}
\usage{
init_rrtools(path, ...)
}
\arguments{
\item{path}{location to create new package. The last component of the path will be used as the package name}

\item{...}{list of description values to override default values or add additional values}
}
\description{
Working on it...
}
