% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{knit_and_update_desc_imports}
\alias{knit_and_update_desc_imports}
\title{Update the package DESCRIPTION with packages used in the Rmd}
\usage{
knit_and_update_desc_imports(inputFile, encoding)
}
\arguments{
\item{file}{path to the Rmd}
}
\value{
nothing
}
\description{
This looks into the Rmd file and identifies packages used
in the code there. Then it updates the DESCRIPTION file of the custom
compendium package that contains the Rmd. It adds the names of packages
used in the Rmd to the Imports field of the DESCRIPTION. It is intended
to be used as a knitr hook with something like this in the paper.Rmd
YAML
\code{knit: (function(inputFile, encoding) { knit_and_update_desc_imports(inputFile, encoding) })}
}
\examples{
\dontrun{
---
knit: (function(inputFile, encoding) { knit_and_update_desc_imports(inputFile, encoding) }
---
}
}
