% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krig_SNMF_gdm.r
\name{krig_SNMF_gdm}
\alias{krig_SNMF_gdm}
\title{Performs spatial krig of population ancestry coefficients
for prediction using R package gdm}
\usage{
krig_SNMF_gdm(gdm_dir, pixels_per_degree = 20, buff = 0.5,
  krig_lambda = NULL)
}
\arguments{
\item{gdm_dir}{-- directory where gdm was performed}

\item{pixels_per_degree}{-- spatial resolution}

\item{buff}{-- spatial buffer}

\item{krig_lambda}{-- parameter for krig, if NULL, uses default}
}
\value{
a list of rasters for use in predicting gdm on krigged ancestry coefficients
}
\description{
Performs spatial krig of population ancestry coefficients
for prediction using R package gdm
}
\examples{
\dontrun{
gdm_files <- dart2gdm(gms, etc)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
