% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_srast.r
\name{make_srast}
\alias{make_srast}
\title{Creates a raster with chosen resolution and buffers, and arbitrary values
Used to create a raster to predict a gdm model onto}
\usage{
make_srast(gdm_dir, pixels_per_degree = 20, buff = 0.5)
}
\arguments{
\item{gdm_dir}{-- directory where gdm was performed}

\item{pixels_per_degree}{-- spatial resolution}

\item{buff}{-- spatial buffer}
}
\value{
a list of rasters for use in predicting gdm on krigged ancestry coefficients
}
\description{
Creates a raster with chosen resolution and buffers, and arbitrary values
Used to create a raster to predict a gdm model onto
}
\examples{
\dontrun{
gdm_files <- dart2gdm(gms, etc)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
