% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipop.Fst.r
\name{multipop.Fst}
\alias{multipop.Fst}
\title{Returns a matrix of spatial distances among populations}
\usage{
multipop.Fst(dart_data, population, basedir, species, dataset)
}
\arguments{
\item{dart_data}{-- a dart meta data object [Required]}

\item{population}{-- a dart meta data object [Required]}
}
\value{
Fst estimates across a nominated set of populations
}
\description{
Input is a dart meta data object
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

