% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openRRstandard.r
\name{openRRstandard}
\alias{openRRstandard}
\title{Write a dart data list object to an R workspace file}
\usage{
openRRstandard(basedir, species, dataset, treatment)
}
\arguments{
\item{species}{-- name of the species in GenuSpec format [required]}

\item{treatment}{-- the treatment string, which is used to find file [required]}

\item{dart_data}{-- dart data list  [required]}

\item{name}{-- arbitrary name for the dataset [required]}
}
\description{
write.dart.data() receives dart data in list format, and reports a series of quality stats.
These are written to RandR/data/species/qual_stats/name*
}
\examples{
write.dart.data(dart_data, RandRbase, "FicuRubi", "DFi16-2154")
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

