% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.pw.Fst.r
\name{population.pw.Fst}
\alias{population.pw.Fst}
\title{Returns a matrix of spatial distances among populations}
\usage{
population.pw.Fst(dart_data, population, basedir, species, dataset)
}
\arguments{
\item{dart_data}{-- a dart meta data object [Required]}
}
\value{
a matrix containing spatial distances between populations
}
\description{
Input is a dart meta data object
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

