% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gdm.r
\name{predict_gdm}
\alias{predict_gdm}
\title{Clip a raster to a convex hull surrounding a set of points}
\usage{
predict_gdm(gdm_dir, gdm_model, srast = NULL, Q = FALSE, qrast = NULL,
  E = FALSE, erast = NULL, s1_point = NULL, s2_point = NULL,
  point_to_point = FALSE, point_to_all = FALSE, all_to_all = FALSE)
}
\arguments{
\item{gdm_dir}{-- the directory containing the gdm}

\item{gdm_model}{-- the gdm model to use in predictions}

\item{srast}{-- a spatial raster for gdm prediction}

\item{Q}{-- Q-matrix (ancestry coefficients) used in gdm}

\item{qrast}{-- if Q, a raster of Q matrix values}

\item{E}{-- Environmental variables used in the model}

\item{erast}{-- if E, a raster of environmental values}

\item{s1_point}{-- a location, c(long, lat), for predicting dissimilarity}

\item{s2_point}{-- a location, c(long, lat), for predicting dissimilarity}

\item{point_to_point}{-- if TRUE, model dissimilarity between s1 and s2}

\item{point_to_all}{-- if TRUE, model dissimilarity between s1 and all points in srast}
}
\value{
file name
}
\description{
Clip a raster to a convex hull surrounding a set of points
}
\examples{
\dontrun{
new_raster <- clip_raster_to_points(old_raster, location_file="locations.txt")
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
