% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_SNMF_for_gdm.r
\name{process_SNMF_for_gdm}
\alias{process_SNMF_for_gdm}
\title{Processes SNMF output for use as explanatory variables in GDM}
\usage{
process_SNMF_for_gdm(basedir, species, dataset, treatment, pop, Ksel)
}
\arguments{
\item{basedir}{-- name of the base directory for R&R}

\item{species}{-- species name}

\item{dataset}{-- dataset name}

\item{treatment}{}

\item{pop}{-- a vector of population assignments}

\item{Ksel}{-- the selected number of ancestral populations}
}
\value{
file name
}
\description{
Processes SNMF output for use as explanatory variables in GDM
}
\examples{
\dontrun{
gdm_files <- SNMF2gdm(basedir, species, dataset, treatment, process=TRUE, Ksel=2)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
