% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.RR.data.base.r
\name{query.RR.data.base}
\alias{query.RR.data.base}
\title{Read metadata for a DArT dataset from RR database}
\usage{
query.RR.data.base(dart_data, basedir, species, dataset, RR_db_file,
  overwrite_meta = TRUE)
}
\arguments{
\item{datafile}{-- name of xls file containing the meta data [required]}
}
\value{
-- a dart meta data object, consisting of sample names
}
\description{
query.RR.data.base() opens a meta data file associated with a dart dataset
}
\author{
Jason Bragg (jasongbragg@gmail.com)
meta_data_file <- query.RR.data.base(dart_data, basedir, species, dataset, RR_db_file, overwrite_meta=TRUE)
}
