% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dart.xls.onerow.r
\name{read.dart.xls.onerow}
\alias{read.dart.xls.onerow}
\title{Import DArT genotypes and store in a list}
\usage{
read.dart.xls.onerow(basedir, species, dataset, topskip, nmetavar, nas = "-",
  altcount = TRUE, euchits = FALSE)
}
\arguments{
\item{basedir}{-- name of csv file containing the DArT data [required]}

\item{species}{-- name of the species in 'GenuSpec' format [required]}

\item{dataset}{-- dataset name assigned by DArT, appears in xlsx file name [required]}

\item{topskip}{-- number of rows to skip before the header row (containing the specimen identities) [required]}

\item{nas}{-- missing data character [default "-"]}

\item{altcount}{-- if TRUE, change genotype encoding to the count of 'alternate' allele  [default TRUE]}

\item{euchits}{-- if TRUE, gets information on BLAST hits in Eucalylptus genome  [default TRUE]}

\item{nmetvar}{-- number of columns containing the locus metadata (e.g. CloneID, Reproducibility) [required]}
}
\value{
A DArT data list, genotype matrix, locus names (CloneID), sample names, Reproducibility scores (RepAvg)
}
\description{
read.dart.xls.onerow() opens the onerow data format supplied by DArT
it is assumed this will be the fourth sheet in an xlsx file 
It skips the first n=topskip lines, and assumes the next line contains  
sample labels followed immediately by genotype data.
}
\details{
It reads the genotype calls in dart format (encoding=dart) and
can covert these to counts of the alternate allele (encoding=altcount)

The data matrix, locus names, metadata, sample names, are written to a dart data
object. This is simply a list containing some specified fields
}
\examples{
\dontrun{
dart_data            <- read.dart.xls.onerow(RandRbase,species,dataset,topskip, nmetavar, euchits=TRUE)
} 
}
\author{
Jason Bragg (jasongbragg@gmail.com), based on template authored by Arthur Georges (glbugs@aerg.canberra.edu.au)
}
