% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.meta.data.r
\name{read.meta.data}
\alias{read.meta.data}
\title{Import metadata for a DArT dataset}
\usage{
read.meta.data(dart_data, basedir, species, dataset, fields = 1, nas = "-")
}
\arguments{
\item{nas}{-- missing data character [default "-"]}

\item{datafile}{-- name of xls file containing the meta data [required]}
}
\value{
-- a dart meta data object, consisting of sample names
}
\description{
read.meta.data() opens a meta data file associated with a dart dataset
}
\examples{
meta_data <- read.dart.meta(datafile=, nas="-")
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

