% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.meta.info.r
\name{read.meta.info}
\alias{read.meta.info}
\title{Import metadata for a DArT dataset}
\usage{
read.meta.info(dart_data, basedir, species, dataset, version = "db",
  fields = 1)
}
\arguments{
\item{datafile}{-- name of xls file containing the meta data [required]}

\item{nas}{-- missing data character [default "-"]}
}
\value{
-- a dart meta data object, consisting of sample names
}
\description{
read.meta.data() opens a meta data file associated with a dart dataset
}
\examples{
meta_data <- read.dart.meta(datafile=, nas="-")
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
