% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.duplicate.samples.r
\name{remove.duplicate.samples}
\alias{remove.duplicate.samples}
\title{Identify duplicate samples and remove one}
\usage{
remove.duplicate.samples(dart_data, least_missing = TRUE,
  remove_fixed_loci = TRUE)
}
\arguments{
\item{dart_data}{-- dart data list  [required]}

\item{least_missing}{-- if TRUE, by least missing}

\item{seed}{-- random seed}
}
\description{
remove.duplicate.samples() receives data in list format. 
Keeps sample missing least data, and thereafter, by precedence in file
}
\examples{
remove.duplicate.samples(dart_data, least_missing=TRUE, remove_fixed_loci=TRUE)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

