% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.fixed.snps.r
\name{remove.fixed.snps}
\alias{remove.fixed.snps}
\title{Remove fixed SNP sites from a DArT data object}
\usage{
remove.fixed.snps(dart_data)
}
\arguments{
\item{dart_data}{-- dart data list  [required]}
}
\description{
remove.fixed.snps() receives dart data in list format, and removes SNP sites that are fixed
}
\examples{
remove.fixed.snps(dart_data)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
