% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.poor.quality.snps.r
\name{remove.poor.quality.snps}
\alias{remove.poor.quality.snps}
\title{Removes SNP sites from a DArT data object subject to quality criteria}
\usage{
remove.poor.quality.snps(dart_data, min_repro = 0.975, max_missing = 0.25)
}
\arguments{
\item{dart_data}{-- dart data list  [required]}

\item{min_repro}{-- minimum value of RepAvg}

\item{max_missing}{-- maximum proportion of missing data [required]}
}
\description{
remove.poor.quality.snps() receives dart data in list format, 
and removes SNP sites that fall below presence and RepAvg thresholds
}
\examples{
remove.poor.quality.snps(dart_data, min_repro=0.975, max_missing=0.25)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

