% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.dart.qc.stats.r
\name{report.dart.qc.stats}
\alias{report.dart.qc.stats}
\title{Report DArT quality statistics}
\usage{
report.dart.qc.stats(dart_data, basedir, species, dataset,
  threshold_missing_loci = 0.5)
}
\arguments{
\item{dart_data}{-- dart data list  [required]}

\item{species}{-- name of the species in GenuSpec format [required]}

\item{dataset}{-- arbitrary name for the dataset [required]}
}
\description{
report.dart.qc.stats() receives dart data in list format, and reports a series of quality stats.
These are written to basedir/species/qual_stats/name*
}
\examples{
report.dart.qc.stats(dart_data, RandRdir, "FicuRubi", "DFi16-2154")
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
