% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.species.params.r
\name{RR.species.params}
\alias{RR.species.params}
\title{Read from a table of RR genotype datasets, get species-specific parameters}
\usage{
RR.species.params(paramdir, species_file, target_species = "all")
}
\arguments{
\item{paramdir}{-- name of directory containing a file with a standard parameter set [required]}

\item{species_file}{-- name of the file containing information about the genotype data [required]}
}
\value{
-- A list of parameters, which can be used in various data filtering functions
}
\description{
Read from a table of RR genotype datasets, get species-specific parameters
}
\examples{
\dontrun{
p <- RR.species.params(RandRparam, species_file)
} 
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
