% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.standard.filters.r
\name{RR.standard.filters}
\alias{RR.standard.filters}
\title{Read genotype filtering parameters from a file}
\usage{
RR.standard.filters(paramdir, param_file)
}
\arguments{
\item{paramdir}{-- name of directory containing a file with a standard parameter set [required]}

\item{param_file}{-- name of the file assigning values to parameters [required]}
}
\value{
-- A list of parameters, which can be used in various data filtering functions
}
\description{
Read genotype filtering parameters from a file
}
\examples{
\dontrun{
p <- RR.standard.filters(RandRparam, param_file)
} 
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

