% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrtools-package.R
\docType{package}
\name{rrtools-package}
\alias{rrtools}
\alias{rrtools-package}
\title{rrtools: Creates a Reproducible Research Compendium}
\description{
Instructions, templates, and functions for making a basic 
    compendium suitable for doing reproducible research with R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benmarwick/rrtools}
  \item Report bugs at \url{https://github.com/benmarwick/rrtools/issues}
}

}
\author{
\strong{Maintainer}: Ben Marwick \email{benmarwick@gmail.com}

Other contributors:
\itemize{
  \item Daniel Knitter [contributor]
  \item Patrick Kennedy [contributor]
  \item Nils Muller-Scheessel [contributor]
  \item Martin Hinz [contributor]
  \item Clemens Schmid [contributor]
  \item Ricarda Braun [contributor]
  \item Wojciech Francuzik [contributor]
  \item Charles Gray [contributor]
  \item Joseph de la Torre Dwyer [contributor]
}

}
\keyword{internal}
