% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_BEDASSLE.r
\name{run_BEDASSLE}
\alias{run_BEDASSLE}
\title{Write a dart data list object to an R workspace file}
\usage{
run_BEDASSLE(bd)
}
\arguments{
\item{bd_par}{-- list of BEDASSLE params  [required]}
}
\description{
run_sunder() receives dart data in list format, and reports a series of quality stats.
These are written to RandR/data/species/qual_stats/name*
}
\examples{
run_BEDASSLE(bd_par)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
