% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diveRsity.r
\name{run_diveRsity}
\alias{run_diveRsity}
\title{Write a dart data list object to an R workspace file}
\usage{
run_diveRsity(gp_file)
}
\arguments{
\item{gp_file}{-- location of a genepop file [required]}
}
\description{
run_sunder() receives dart data in list format, and reports a series of quality stats.
These are written to RandR/data/species/qual_stats/name*
}
\examples{
run_diveRsity(gp_file)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

