% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SpaceMix.r
\name{run_SpaceMix}
\alias{run_SpaceMix}
\title{Write a dart data list object to an R workspace file}
\usage{
run_SpaceMix(sm)
}
\arguments{
\item{sm}{-- list of SpaceMix data and params  [required]}
}
\description{
run_sunder() receives dart data in list format, and reports a series of quality stats.
These are written to RandR/data/species/qual_stats/name*
}
\examples{
run_SpaceMix(sm)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
