% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sunder.r
\name{run_sunder}
\alias{run_sunder}
\title{Write a dart data list object to an R workspace file}
\usage{
run_sunder(s, basedir, species, dataset, treatment)
}
\arguments{
\item{species}{-- name of the species in GenuSpec format [required]}

\item{treatment}{-- the treatment string, which is used to find file [required]}

\item{sun_par}{-- list of sunder params  [required]}

\item{name}{-- arbitrary name for the dataset [required]}
}
\description{
run_sunder() receives dart data in list format, and reports a series of quality stats.
These are written to RandR/data/species/qual_stats/name*
}
\examples{
run_sunder(dart_data, RandRbase, "FicuRubi", "DFi16-2154")
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
