% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.one.snp.per.locus.random.r
\name{sample.one.snp.per.locus.random}
\alias{sample.one.snp.per.locus.random}
\title{Sample one SNP per locus at random}
\usage{
sample.one.snp.per.locus.random(dart_data, seed = 12345)
}
\arguments{
\item{dart_data}{-- dart data list  [required]}

\item{seed}{-- random seed}
}
\description{
sample.one.snp.per.locus.random() receives data in list format, and samples one snp per locus randomly
A new dart data list object is written, with the treatment field updated to "1SNPperClone"
}
\examples{
sample.one.snp.per.locus.random(dart_data, seed=12345)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
