% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.snp.random.r
\name{sample.snp.random}
\alias{sample.snp.random}
\title{Sample one SNP per locus at random}
\usage{
sample.snp.random(dart_data, number = 5000, seed = 12345)
}
\arguments{
\item{dart_data}{-- dart data list  [required]}

\item{number}{-- number of random snps to select}

\item{seed}{-- random seed}
}
\description{
sample.snp.random() receives data in list format, and samples a nominated number
of snp sites at random, with the treatment field updated to "downSamp"
}
\examples{
sample.snp.random(dart_data, seed=12345)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
