% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.eucalypt.genome.hits.r
\name{save.eucalypt.genome.hits}
\alias{save.eucalypt.genome.hits}
\title{Save information on DArT blast hits to eucalypt genome}
\usage{
save.eucalypt.genome.hits(x, basedir, species, dataset)
}
\arguments{
\item{x}{-- matrix of DArT data, supplied by DArT, as exctracted from excel file [required]}

\item{basedir}{-- name of csv file containing the DArT data [required]}

\item{species}{-- name of the species in 'GenuSpec' format [required]}

\item{dataset}{-- dataset name assigned by DArT, appears in xlsx file name [required]}
}
\value{
A statement about whether list consisting of a matrix of genotypes, locus names (CloneID), sample names, Reproducibility scores (RepAvg)
}
\description{
save.eucalypt.genome.hits() saves information on DArT hits to eucalypt genome
This function is expected to be called during initial extraction of data
by the function read.dart.xls.onerow
}
\examples{
dart_data <- read.dart.onerow(datafile="Report-DFi16-2154_oneline.csv", topskip=6, nmetavar=38, nas="-")
}
\author{
Jason Bragg (jasongbragg@gmail.com), based on template authored by Arthur Georges (glbugs@aerg.canberra.edu.au)
}

