% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNMF2gdm.r
\name{SNMF2gdm}
\alias{SNMF2gdm}
\title{Prepares SNMF Q matrices to act as explanator variables in GDM}
\usage{
SNMF2gdm(basedir, species, dataset, treatment, process = FALSE, pop = NULL,
  Ksel = NULL)
}
\arguments{
\item{basedir}{-- name of the base directory for R&R}

\item{species}{-- species name}

\item{dataset}{-- dataset name}

\item{treatment}{}

\item{process}{[TRUE / FALSE] -- indicates whether to process Q matrix, 
or if it has been processed previously}
}
\value{
file name
}
\description{
Prepares SNMF Q matrices to act as explanator variables in GDM
}
\examples{
\dontrun{
gdm_files <- SNMF2gdm(basedir, species, dataset, treatment, process=TRUE, Ksel=2)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
