% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{use_compendium}
\alias{use_compendium}
\alias{add_travis}
\title{Creates an R package suitable to use as a research compendium

This is devtools::create() with an additional step to either start the project in RStudio, or set the working directory to the pkg location, if not using RStudio}
\usage{
use_compendium(path, description = getOption("devtools.desc"),
  check = FALSE, rstudio = TRUE, quiet = FALSE)
}
\arguments{
\item{browse}{open a browser window to enable Travis builds for the package
automatically.}
}
