% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_git.R
\name{use_git_quietly}
\alias{use_git_quietly}
\title{Initialise a git repository without asking questions}
\usage{
use_git_quietly(message = "Initial commit")
}
\arguments{
\item{message}{Message to use for first commit.}
}
\description{
From usethis, modified to be non-interactive.
`use_git_quietly()` initialises a Git repository and adds important files to
`.gitignore`. If user consents, it also makes an initial commit.
}
\examples{
\dontrun{
use_git_quietly()
}
}
\concept{git helpers}
