% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_use_gitlab.R
\name{use_gitlab}
\alias{use_gitlab}
\title{Title}
\usage{
use_gitlab(pkg = ".", auth_token = "xxxx", rocker = "verse",
  rmd_to_knit = "path_to_rmd")
}
\arguments{
\item{pkg}{file location of package}

\item{auth_token}{personal access token}

\item{rocker}{name the rocker image}

\item{rmd_to_knit}{the path to the .Rmd}
}
\description{
Title
}
