% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{use_travis}
\alias{use_travis}
\alias{add_travis}
\title{Add a travis config file

This differs from devtools by directing travis to build the Docker container and push the successful result to Docker Hub.}
\usage{
use_travis(pkg = ".", browse = interactive())
}
\arguments{
\item{browse}{open a browser window to enable Travis builds for the package
automatically.}
}
