% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.dart.data.r
\name{write.dart.data}
\alias{write.dart.data}
\title{Write a dart data list object to an R workspace file}
\usage{
write.dart.data(dart_data, basedir, species, dataset, meta = FALSE)
}
\arguments{
\item{dart_data}{-- dart data list  [required]}

\item{species}{-- name of the species in GenuSpec format [required]}

\item{name}{-- arbitrary name for the dataset [required]}
}
\description{
write.dart.data() receives dart data in list format, and reports a series of quality stats.
These are written to RandR/data/species/qual_stats/name*
}
\examples{
write.dart.data(dart_data, RandRbase, "FicuRubi", "DFi16-2154")
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
