% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_manage_dependencies.R
\name{add_dependencies_to_description}
\alias{add_dependencies_to_description}
\title{Searches for external packages and adds them to the Imports field in the description}
\usage{
add_dependencies_to_description(
  path = getwd(),
  description_file = "DESCRIPTION",
  just_packages = FALSE
)
}
\arguments{
\item{path}{location of individual file or directory where to search for scripts.}

\item{description_file}{location of description file to be updated.}

\item{just_packages}{just give back a character vector of the found packages.}
}
\description{
Scans script files (.R, .Rmd, .Rnw, .Rpres, etc.) for external package dependencies indicated by
\code{library()}, \code{require()} or \code{::} and adds those packages to the Imports field in
the package DESCRIPTION.
}
