% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{use_analysis}
\alias{use_analysis}
\title{Adds and analysis directory (and sub-directories), and an Rmd file ready to write

This will create \file{analysis/paper.Rmd}, \file{analysis/references.bib}
and several others, and add \pkg{bookdown} to the imported packages.}
\usage{
use_analysis(pkg = ".", template = "paper.Rmd", data = list())
}
