% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{use_circleci}
\alias{use_circleci}
\title{Add a circleci config file}
\usage{
use_circleci(pkg = ".", browse = interactive(), docker_hub = TRUE)
}
\arguments{
\item{pkg}{defaults to the package in the current working directory}

\item{browse}{open a browser window to enable Circle-CI builds for the package automatically}
}
\description{
This will build the Docker container on the Circle-CI service.
The advantage of Circle-CI over Travis is that Circle-CI will freely work with
private GitHub repositories. Only the paid service from Travis will work
with private GitHub repositories. Before using this function you need
to create an account with Circle-CI, using youg GitHub account. If you want
Circle-CI to run on a private GitHub repo, make sure you give Circle-CI
access to 'all repos' when you log in with your GitHub credentials.
}
