% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_use_compendium.R
\name{use_compendium}
\alias{use_compendium}
\title{Creates an R package suitable to use as a research compendium, and
switches to the working directory of this new package, ready to work}
\usage{
use_compendium(
  path = getwd(),
  fields = getOption("usethis.description"),
  rstudio = rstudioapi::isAvailable(),
  open = TRUE,
  quiet = FALSE,
  simple = TRUE,
  welcome_message = TRUE
)
}
\arguments{
\item{path}{location to create new package. The last component of the path will be used as the package name}

\item{fields}{list of description values to override default values or add additional values}

\item{rstudio}{create an RStudio project file? (with \code{usethis::use_rstudio})}

\item{open}{if TRUE and in RStudio, the new project is opened in a new instance.
If TRUE and not in RStudio, the working directory is set to the new project}

\item{quiet}{if FALSE, the default, prints informative messages}

\item{simple}{if TRUE, the default, the R/ directory is not created, because it's not necessary
for many if not most research repositories}

\item{welcome_message}{if TRUE, rstudio, open and not quiet, then the .Rprofile file in the
newly created package is prepopulated with a welcome message.}
}
\description{
This is usethis::create_package() with some additional messages to simplify the transition into the new project setting
}
