% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{use_dockerfile}
\alias{use_dockerfile}
\title{Add a Dockerfile

This will create a basic \file{Dockerfile} based on rocker/verse}
\usage{
use_dockerfile(pkg = ".", rocker = "verse")
}
\arguments{
\item{rocker}{chr, the rocker image to base this container on}
}
