% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_use_dockerfile.R
\name{use_dockerfile}
\alias{use_dockerfile}
\title{Add a Dockerfile}
\usage{
use_dockerfile(
  pkg = ".",
  rocker = "verse",
  qmd_to_knit = "path_to_qmd",
  use_gh_action = TRUE
)
}
\arguments{
\item{pkg}{defaults to the package in the current working directory}

\item{rocker}{chr, the rocker image to base this container on}

\item{qmd_to_knit, }{chr, path to the qmd file to render in the Docker
container, relative to the top level of the compendium
(i.e. "analysis/paper/paper.qmd"). There's no need to specify this if your qmd
to render is at "analysis/paper/paper.qmd", "vignettes/paper/paper.qmd" or
"inst/paper/paper.qmd". If you have a custom directory structure, and a custom
file name for the qmd file, you can specify that file path and name here so
Docker can find the file to render in the container.B}
}
\description{
This will create a basic \file{Dockerfile} based on rocker/verse
}
